pyproject.toml
[project]
name = "test-analysis-recommendation-vertex-ai-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.5",
    "fastapi>=0.115.6",
    "google-cloud-aiplatform>=1.36.0",
    "google-cloud-storage>=2.8.0",
    "google-genai>=1.53.0",
    "google-generativeai>=0.3.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "ulid-py>=1.1.0",
    "uvicorn>=0.32.1",
    "vertexai>=0.0.1",
]
